requirements-dev.txt
# httpx mocking for Helium-10 fee-cron tests
Jinja2==3.1.6
alembic==1.17.0
apscheduler==3.10.4  # async scheduler for freight cron
asgi-correlation-id==4.3.1
asyncpg==0.30.0
black==24.8.0  # formatter used in CI step "black --check ."
boto3==1.40.21
redis==6.*
minio==7.2.18
vulture==2.10
celery==5.5.3
fastapi==0.116.1
fastapi[all]==0.116.1
fastapi-limiter==0.1.6
httpx>=0.27,<0.28
imapclient==3.0.1
itsdangerous
mypy==1.*
mypy-extensions==1.1.0
numpy<2
openpyxl==3.1.5
pandas==2.3.2
psycopg2-binary==2.9.10
psycopg[binary]==3.2.11
pydantic>=2.6
pydantic-settings>=2.1
setuptools
pytest-asyncio==1.1.0
pytest-cov==7.0.0
pytest-postgresql==7.0.2
python-dotenv
python-multipart
python-telegram-bot==22.5
requests==2.32.5
respx==0.22.0
ruff==0.7.0  # static linter used in CI step "ruff check ."
sqlalchemy==2.0.43
structlog==24.1.0
testcontainers[postgres]==4.12.0
types-requests==2.32.4.20250809
docspec-python==2.2.2
pydoc-markdown==4.8.2
pandera==0.26.*
